// prisma/enum.ts
export enum ROLES_ENUM {
  PLATFORM_USER = 'PLATFORM_USER',
  BUSINESS_USER = 'BUSINESS_USER',
  PLATFORM = 'PLATFORM',
  CLIENT = 'CLIENT',
  COMPANY = 'COMPANY',
  BRANCH = 'BRANCH',
  STAFF = 'STAFF',
  AGENCY = 'AGENCY',
}

export enum ADDRESS_TYPE_ENUM {
  USER_HOME = 'user_home',
  USER_WORK = 'user_work',
  BUSINESS_PRIMARY = 'business_primary',
  BUSINESS_BRANCH = 'business_branch',
  PROFI_SERVICE_AREA = 'profi_service_area',
  ORDER_DELIVERY = 'order_delivery',
  TRAINING_VENUE = 'training_venue',
  INSTRUCTOR_LOCATION = 'instructor_location',
  BILLING_ADDRESS = 'billing_address',
  BANK_BRANCH = 'bank_branch',
  MANUFACTURER_LOCATION = 'manufacturer_location',
  SUPPLIER_LOCATION = 'supplier_location',
  WAREHOUSE_LOCATION = 'warehouse_location',
  STORAGE_LOCATION = 'storage_location',
  CREATED_CLIENT = 'created_client',
}

export enum FILE_ENTITY_TYPE_ENUM {
  USER_AVATAR = 'user_avatar',
  BUSINESS_LOGO = 'business_logo',
}